Expose awsApi, formatTemplateString functions for elasticsearch
