+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "cloudli",
-  "version": "1.0.0",
+  "version": "1.1.0",
   "description": "Command line utilities for Google Cloud Firestore and Amazon Elasticsearch",
   "author": "Duane Wood",
   "license": "MIT",
