+++ src/elasticsearch/commands/utils.js
@@ -1,3 +1,6 @@
+const get = require('lodash.get')
+const isPlainObject = require('lodash.isplainobject')
+
 /**
  * Gets the list of IndexConfig objects based on command line parameters.
  * If index is null or undefined, defaults to all indices in config.
@@ -42,6 +45,27 @@ function getIndexConfigsFromParams(index, options, config) {
   return indices
 }
 
+/**
+ * Formats a template string containing substitution parameters in the form ${name}
+ *
+ * @param {string} template the template string containing substitution parameters in the form ${name}.
+ *                          name may be a nested name.  For example, ${item[0].product.name}
+ * @param {object} vars the object to use for substitution.  name is resolved within vars.
+ *                          If any of the substitution values is an object, JSON.stringify will be
+ *                          used for the substitution.
+ * @return {string} the formatted string.
+ */
+const formatTemplateString = (template, vars) =>
+  template.replace(/\${(.*?)}/g, (_, v) => {
+    const value = get(vars, v, v)
+    if (isPlainObject(value)) {
+      return JSON.stringify(value, null, 2)
+    } else {
+      return value
+    }
+  })
+
 module.exports = {
-  getIndexConfigsFromParams
+  getIndexConfigsFromParams,
+  formatTemplateString
 }
